Revert "Rename ECSCredentials to RemoteCredentials and derive new ECSCredentials from it (#1806)"
This reverts commit 8a61fe5f830ca9b9e10e05e5c6859f1c4483b174.
