--- .changes/next-release/feature-credentials_provider-683a1b9e.json
@@ -1,5 +1,0 @@
-{
-  "type": "feature",
-  "category": "credentials_provider",
-  "description": "Add general purpose RemoteCredentials and derive ECSCredentials from it"
-}
