--- lib/credentials/remote_credentials.js
@@ -1,236 +1,0 @@
-var AWS = require('../core'),
-  FULL_URI_UNRESTRICTED_PROTOCOLS = ['https:'],
-  FULL_URI_ALLOWED_PROTOCOLS = ['http:', 'https:'],
-  FULL_URI_ALLOWED_HOSTNAMES = ['localhost', '127.0.0.1'],
-  RELATIVE_URI_HOST = '169.254.170.2';
-
-/**
- * Represents credentials received from specified URI.
- *
- * This class will request refreshable credentials from the relative URI
- * specified by the AWS_REMOTE_CREDENTIALS_RELATIVE_URI or the
- * AWS_REMOTE_CREDENTIALS_FULL_URI environment variable. If valid credentials
- * are returned in the response, these will be used with zero configuration.
- *
- * This credentials class will by default timeout after 1 second of inactivity
- * and retry 3 times.
- * If your requests to the relative URI are timing out, you can increase
- * the value by configuring them directly:
- *
- * ```javascript
- * AWS.config.credentials = new AWS.RemoteCredentials({
- *   httpOptions: { timeout: 5000 }, // 5 second timeout
- *   maxRetries: 10, // retry 10 times
- *   retryDelayOptions: { base: 200 } // see AWS.Config for information
- * });
- * ```
- *
- * @see AWS.Config.retryDelayOptions
- *
- * @!macro nobrowser
- */
-AWS.RemoteCredentials = AWS.util.inherit(AWS.Credentials, {
-  constructor: function RemoteCredentials(options) {
-    AWS.Credentials.call(this);
-    options = options ? AWS.util.copy(options) : {};
-    if (!options.httpOptions) options.httpOptions = {};
-    options.httpOptions = AWS.util.merge(
-      this.httpOptions, options.httpOptions);
-    AWS.util.update(this, options);
-  },
-
-  /**
-   * @api private
-   */
-  providerName: 'RemoteCredentials',
-
-  /**
-   * @api private
-   */
-  envRelativeUri: 'AWS_REMOTE_CREDENTIALS_RELATIVE_URI',
-
-  /**
-   * @api private
-   */
-  envFullUri: 'AWS_REMOTE_CREDENTIALS_FULL_URI',
-
-  /**
-   * @api private
-   */
-  envAuthToken: 'AWS_REMOTE_AUTHORIZATION_TOKEN',
-
-  /**
-   * @api private
-   */
-  httpOptions: { timeout: 1000 },
-
-  /**
-   * @api private
-   */
-  maxRetries: 3,
-
-  /**
-   * @api private
-   */
-  isConfiguredForRemoteCredentials: function isConfiguredForRemoteCredentials() {
-    return Boolean(
-        process &&
-        process.env &&
-        (process.env[this.envRelativeUri] || process.env[this.envFullUri])
-    );
-  },
-
-  /**
-   * @api private
-   */
-  getRemoteFullUri: function getRemoteFullUri() {
-    if (process && process.env) {
-      var relative = process.env[this.envRelativeUri],
-          full = process.env[this.envFullUri];
-      if (relative) {
-        return 'http://' + RELATIVE_URI_HOST + relative;
-      } else if (full) {
-        var parsed = AWS.util.urlParse(full);
-        if (FULL_URI_ALLOWED_PROTOCOLS.indexOf(parsed.protocol) < 0) {
-          throw AWS.util.error(
-            new Error('Unsupported protocol:  AWS.' + this.providerName + ' supports '
-              + FULL_URI_ALLOWED_PROTOCOLS.join(',') + ' only; '
-              + parsed.protocol + ' requested.'),
-            { code: this.providerName + 'ProviderFailure' }
-          );
-        }
-
-        if (FULL_URI_UNRESTRICTED_PROTOCOLS.indexOf(parsed.protocol) < 0 &&
-            FULL_URI_ALLOWED_HOSTNAMES.indexOf(parsed.hostname) < 0) {
-          throw AWS.util.error(
-            new Error('Unsupported hostname: AWS.' + this.providerName + ' only supports '
-              + FULL_URI_ALLOWED_HOSTNAMES.join(',') + ' for ' + parsed.protocol + '; '
-              + parsed.protocol + parsed.hostname + ' requested.'),
-            { code: this.providerName + 'ProviderFailure' }
-          );
-        }
-
-        return full;
-      }
-    }
-  },
-
-  /**
-   * @api private
-   */
-  getRemoteAuthToken: function getRemoteAuthToken() {
-    if (process && process.env && process.env[this.envFullUri]) {
-      return process.env[this.envAuthToken];
-    }
-  },
-
-  /**
-   * @api private
-   */
-  credsFormatIsValid: function credsFormatIsValid(credData) {
-    return (!!credData.accessKeyId && !!credData.secretAccessKey &&
-      !!credData.sessionToken && !!credData.expireTime);
-  },
-
-  /**
-   * @api private
-   */
-  formatCreds: function formatCreds(credData) {
-    if (!!credData.credentials) {
-      credData = credData.credentials;
-    }
-
-    return {
-      expired: false,
-      accessKeyId: credData.accessKeyId || credData.AccessKeyId,
-      secretAccessKey: credData.secretAccessKey || credData.SecretAccessKey,
-      sessionToken: credData.sessionToken || credData.Token,
-      expireTime: new Date(credData.expiration || credData.Expiration)
-    };
-  },
-
-  /**
-   * @api private
-   */
-  request: function request(url, callback) {
-    var httpRequest = new AWS.HttpRequest(url);
-    httpRequest.method = 'GET';
-    httpRequest.headers.Accept = 'application/json';
-    var token = this.getRemoteAuthToken();
-    if (token) {
-      httpRequest.headers.Authorization = token;
-    }
-    AWS.util.handleRequestWithRetries(httpRequest, this, callback);
-  },
-
-  /**
-   * @api private
-   */
-  refreshQueue: [],
-
-  /**
-   * Loads the credentials from the relative URI specified by container
-   *
-   * @callback callback function(err)
-   *   Called when the request to the relative URI responds (or fails). When
-   *   this callback is called with no error, it means that the credentials
-   *   information has been loaded into the object (as the `accessKeyId`,
-   *   `secretAccessKey`, `sessionToken`, and `expireTime` properties).
-   *   @param err [Error] if an error occurred, this value will be filled
-   * @see get
-   */
-  refresh: function refresh(callback) {
-    var self = this;
-    var refreshQueue = self.refreshQueue;
-    if (!callback) callback = function(err) { if (err) throw err; };
-    refreshQueue.push({
-      provider: self,
-      errCallback: callback
-    });
-    if (refreshQueue.length > 1) { return; }
-
-    function callbacks(err, creds) {
-      var call, cb;
-      while ((call = refreshQueue.shift()) !== undefined) {
-        cb = call.errCallback;
-        if (!err) AWS.util.update(call.provider, creds);
-        cb(err);
-      }
-    }
-
-    if (process === undefined) {
-      callbacks(AWS.util.error(
-        new Error('No process info available'),
-        { code: this.providerName + 'ProviderFailure' }
-      ));
-      return;
-    }
-    var fullUri = this.getRemoteFullUri();
-    if (fullUri === undefined) {
-      callbacks(AWS.util.error(
-        new Error('Variable ' + this.envRelativeUri + ' or ' + this.envFullUri +
-          ' must be set to use AWS.' + this.providerName + '.'),
-        { code: this.providerName + 'ProviderFailure' }
-      ));
-      return;
-    }
-
-    this.request(fullUri, function(err, data) {
-      if (!err) {
-        try {
-          data = JSON.parse(data);
-          var creds = self.formatCreds(data);
-          if (!self.credsFormatIsValid(creds)) {
-            throw AWS.util.error(
-              new Error('Response data is not in valid format'),
-              { code: this.providerName + 'ProviderFailure' }
-            );
-          }
-        } catch (dataError) {
-          err = dataError;
-        }
-      }
-      callbacks(err, creds);
-    });
-  }
-});
