+++ lib/node_loader.js
@@ -26,7 +26,6 @@ require('./http/node');
 
 // Load custom credential providers
 require('./credentials/ec2_metadata_credentials');
-require('./credentials/remote_credentials');
 require('./credentials/ecs_credentials');
 require('./credentials/environment_credentials');
 require('./credentials/file_system_credentials');
