+++ ts/core.ts
@@ -11,8 +11,6 @@ if (core.CognitoIdentityCredentials) {
 
 } else if (core.EC2MetadataCredentials) {
 
-} else if (core.RemoteCredentials) {
-
 } else if (core.ECSCredentials) {
 
 } else if (core.Endpoint) {
